Update nucleotide count solution
Make the nucleotide key member static so the values are shared between the instances.
Renam the nucleotide validation method to match the style of the rest of the class.
Make tha validation method static so that it can be shared between instances and not required this capturing in lambdas
and when passes to algorithms.
Initialize the mNucleotideCounts throught a initializer list.
Make the  nucleotide counts return a const refference  the to mNucleotideCounts member to avoid
copy.
Make the validation method throw on a invalid nucleotide value to avoid repeating the throw message repeating.